create custom notif. dialog and add confetti
